Incorporate India suggestions
